pyproject.toml
[tool.poetry]
name = "fixprice"
version = "0.1.0"
description = ""
authors = ["Джозеф <joy.life86@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
scrapy = "^2.11.1"
scrapy-splash = "^0.9.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
